Bundle node.js with plugin JAR file
Previously, the plugin only worked in scenarios where the user either
explicitly configured a VM option pointing to the file path of a Node.js
binary, or if the Eclipse installation came included with Node.js via a
`.node` directory that doesn't exist install Eclipse installations.

This issue solves the problem by embedding a Node.js binary for Windows
x64 in the resources of the plugin. The downside of this fix is that the
plugin size is very large on every update even if the Node.js binary is
unchanged. If this is problematic, then we can consider splitting the
Node.js binary into a separate plugin that updates less frequently.