Use "jetbrains" client name instead of "eclipse"
Previously, we send an appropriate Cody agent client name "eclipse".
This was problematic because it meant that the plugin didn't work with
Enterprise server instances that have enabled context filters due to a
check against old versions of clients that may not support context
filters. The server checks have been fixed in https://github.com/sourcegraph/sourcegraph/pull/63855
but we need to wait a few months before being able to send "Eclipse" as
a client name when Enterprise users have completed upgrading to a newer
Sourcegraph version.

The big downside with this solution is that it means all the telemetry
in the plugin indicates that events are coming from JetBrains. We can
fix this problem by extending the Cody agent server protocol to accept
a dedicated setting to override only the HTTP client name for LLM
completion requests. This work is tracked in the issue CODY-2978.