Secrets: move to webview login page
Previously, Eclipse used a custom native "Log in" button that was
awkward to use and wasn't properly connected to the agent process. Now,
we use the new `secrets: 'client-managed'` capability in the
agent protocol that allows Eclipse to get/set/delete secrets.
This means that the webview can just run like in VS Code and
Eclipse only stores the secret.